% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ac_sample1000}
\alias{ac_sample1000}
\title{Sample stats19 dataset}
\format{A data frame}
\description{
Todo: add vehicles and casualties datasets
}
\examples{
\dontrun{
set.seed(2018)
ac_sample1000 = ac[sample(nrow(ac), 1000), ]
usethis::use_data(ac_sample1000, overwrite = TRUE)
}
plot(ac_sample1000$Longitude, ac_sample1000$Latitude)

}
\keyword{datasets}
