% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents_2016_sample}
\alias{accidents_2016_sample}
\title{Sample stats19 data (2016 accidents)}
\format{A data frame}
\description{
Obtained:
dl_stats19(years = "2016", type = "Accide")
acc_2016 = read.csv("/tmp/RtmplOkfPW/dftRoadSafety_Accidents_2016/dftRoadSafety_Accidents_2016.csv")
accidents_2016_sample = head(acc_2016, n = 2)
}
\keyword{datasets}
