% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents_sample}
\alias{accidents_sample}
\alias{accidents_sample_raw}
\title{Sample of stats19 data (2017 accidents)}
\format{A data frame}
\description{
Sample of stats19 data (2017 accidents)
}
\examples{
\dontrun{
# Obtained with:
dl_stats19(year = 2017, type = "Accide")
accidents_2017_raw = read_accidents(year = 2017)
set.seed(350)
sel = sample(nrow(accidents_2017_raw), 3)
accidents_sample_raw = accidents_2017_raw[sel, ]
accidents_sample = format_accidents(accidents_sample_raw)
}
}
\keyword{datasets}
