% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casualties_sample}
\alias{casualties_sample}
\alias{casualties_sample_raw}
\title{Sample of stats19 data (2017 casualties)}
\format{A data frame}
\description{
Sample of stats19 data (2017 casualties)
}
\examples{
\donttest{
nrow(casualties_sample_raw)
casualties_sample_raw
}
}
\keyword{datasets}
