% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{check_input_file}
\alias{check_input_file}
\title{Local helper to be reused.}
\usage{
check_input_file(filename = NULL, type = NULL, data_dir = NULL, year = NULL)
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read.}

\item{type}{One of 'collision', 'casualty', 'Vehicle'.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{year}{Single year for which data are to be read.}
}
\description{
Local helper to be reused.
}
