% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year}
\alias{check_year}
\title{check and convert year argument}
\usage{
check_year(year)
}
\arguments{
\item{year}{Single year for which file is to be downloaded}
}
\description{
check and convert year argument
}
