% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_make}
\alias{clean_make}
\title{Clean vehicle make}
\usage{
clean_make(make)
}
\arguments{
\item{make}{A character vector of vehicle makes}
}
\description{
This function cleans the make of the vehicle.
}
\examples{
clean_make(c("VW", "Mercedez"))
}
