% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_make_model}
\alias{clean_make_model}
\title{Clean vehicle make and model}
\usage{
clean_make_model(generic_make_model)
}
\arguments{
\item{generic_make_model}{A character vector of generic make/model strings}
}
\description{
This function returns a combined cleaned make and model string.
It uses \code{clean_make} and \code{clean_model} to standardize both parts.
}
\examples{
clean_make_model(c("FORD FIESTA", "BMW 3 SERIES"))
}
