% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_model}
\alias{clean_model}
\title{Clean vehicle model}
\usage{
clean_model(model)
}
\arguments{
\item{model}{A character vector of vehicle models}
}
\description{
This function cleans the model of the vehicle.
}
\examples{
clean_model(c("FIESTA", "ka"))
}
