% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_stats19.R
\name{dl_schema}
\alias{dl_schema}
\title{Download stats19 schema}
\usage{
dl_schema(data_dir = tempdir())
}
\description{
This downloads an excel spreadsheet containing variable names and categories
}
\examples{
\dontrun{
dl_schema()
}
}
