% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{domain}
\alias{domain}
\title{Static values
Main location of any changes from upstream
On 22nd Nov 2018 tested both
http://data.dft.gov.uk/road-accidents-safety-data/road-accidents-safety-data/RoadSafetyData_2015.zip
and
http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafety_Accidents_2016
both were available from the s3 url below, so decided to make it the
main domain URL of the stats19}
\format{An object of class \code{character} of length 1.}
\usage{
domain
}
\description{
Static values
Main location of any changes from upstream
On 22nd Nov 2018 tested both
http://data.dft.gov.uk/road-accidents-safety-data/road-accidents-safety-data/RoadSafetyData_2015.zip
and
http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafety_Accidents_2016
both were available from the s3 url below, so decided to make it the
main domain URL of the stats19
}
\keyword{datasets}
