% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_and_unzip}
\alias{download_and_unzip}
\title{Download and unzip given appropriate params}
\usage{
download_and_unzip(fname, zip_url, data_dir = tempdir())
}
\arguments{
\item{fname}{Required zip name also used as destination of csv folder}

\item{zip_url}{Required full path of file to download}

\item{data_dir}{Parent directory of fname}
}
\description{
Downloads dftRoadSafety_2016_Accidents.zip to /tmp/blahDIR
Unzips dftRoadSafety_2016_Accidents.zip to
/tmp/blahDIR/dftRoadSafety_2016_Accidents
and lists what is in it.
}
