% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{extract_make_stats19}
\alias{extract_make_stats19}
\title{Extract vehicle make from generic make/model string}
\usage{
extract_make_stats19(generic_make_model)
}
\arguments{
\item{generic_make_model}{A character vector of generic make/model strings}
}
\description{
This function extracts the make from a generic make/model string, handling multi-word makes.
}
\examples{
extract_make_stats19(c("FORD FIESTA", "LAND ROVER DISCOVERY"))
}
