% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{file_names}
\alias{file_names}
\title{stats19 file names for easy access}
\format{
A named list
}
\description{
URL decoded file names. Currently there are 52 file names
released by the DfT (Department for Transport) and the details include
how these were obtained and would be kept up to date.
}
\note{
These were generated using the script in the
\code{data-raw} directory (\code{misc.Rmd} file).
}
\examples{
head(file_names)
}
\keyword{datasets}
