% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{filter_alongside}
\alias{filter_alongside}
\title{Filter for alongside (sideswipe) collisions}
\usage{
filter_alongside(vehicles)
}
\arguments{
\item{vehicles}{A STATS19 vehicle data frame}
}
\value{
Filtered data frame with pedal cycles in sideswipe collisions
}
\description{
Filter for alongside (sideswipe) collisions
}
