% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{filter_dooring}
\alias{filter_dooring}
\title{Filter for dooring collisions}
\usage{
filter_dooring(vehicles)
}
\arguments{
\item{vehicles}{A STATS19 vehicle data frame}
}
\value{
Filtered data frame with pedal cycles where vehicle was parked/parking
}
\description{
Filter for dooring collisions
}
