% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{filter_failed_give_way}
\alias{filter_failed_give_way}
\title{Filter for failed to give way collisions}
\usage{
filter_failed_give_way(vehicles)
}
\arguments{
\item{vehicles}{A STATS19 vehicle data frame}
}
\value{
Filtered data frame with pedal cycles where vehicle was entering main road
}
\description{
Filter for failed to give way collisions
}
