% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{filter_left_hook}
\alias{filter_left_hook}
\title{Filter for left hook collisions}
\usage{
filter_left_hook(vehicles)
}
\arguments{
\item{vehicles}{A STATS19 vehicle data frame}
}
\value{
Filtered data frame with pedal cycles where the vehicle was turning left
}
\description{
Filter for left hook collisions
}
