% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{filter_right_hook}
\alias{filter_right_hook}
\title{Filter for right hook collisions}
\usage{
filter_right_hook(vehicles)
}
\arguments{
\item{vehicles}{A STATS19 vehicle data frame}
}
\value{
Filtered data frame with pedal cycles where the vehicle was turning right
}
\description{
Filter for right hook collisions
}
