% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_casualties}
\alias{format_casualties}
\title{Format stats19 casualties}
\usage{
format_casualties(x, factorize = FALSE)
}
\arguments{
\item{x}{Data frame created with `read_casualties()`}

\item{factorize}{Should some results be returned as factors? `FALSE` by default}
}
\description{
Format stats19 casualties
}
\section{Details}{

This function formats raw stats19 data
}

\examples{
\dontrun{
dl_stats19(years = 2017, type = "casualties")
x = read_casualties()
casualties = format_casualties(x)
}
}
