% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats19.R
\name{format_stats19_2005_2014_ac}
\alias{format_stats19_2005_2014_ac}
\alias{format_stats19_2015_ac}
\alias{format_stats19_2016_ac}
\title{Format UK 'Stats19' road traffic casualty data}
\usage{
format_stats19_2005_2014_ac(ac, factorize = FALSE)
}
\arguments{
\item{ac}{Data frame representing the raw Stats19 data read-in with `read_csv()`}

\item{factorize}{Should some results be returned as factors? `FALSE` by default}
}
\description{
Format UK 'Stats19' road traffic casualty data
}
\section{Details}{

This is a helper function to format raw stats19 data
}

\examples{
\dontrun{
ac = read_accidents()
sapply(ac, class)
ac_formatted = format_stats19_2005_2014_ac(ac)
sapply(ac_formatted, class)
}
}
