% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_vehicles}
\alias{format_vehicles}
\title{Format stats19 vehicles data}
\usage{
format_vehicles(x, factorize = FALSE)
}
\arguments{
\item{x}{Data frame created with `read_vehicles()`}

\item{factorize}{Should some results be returned as factors? `FALSE` by default}
}
\description{
Format stats19 vehicles data
}
\section{Details}{

This function formats raw stats19 data
}

\examples{
\dontrun{
x = read_vehicles()
sapply(x, class)
table(x$Vehicle_Type)
vehicles = format_vehicles(x)
sapply(vehicles, class)
table(vehicles$vehicle_type)
}
}
