% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mot.R
\name{get_MOT_STATS19}
\alias{get_MOT_STATS19}
\title{Download vehicle data from the DVSA MOT API using VRM.}
\usage{
get_MOT_STATS19(VRMs, apikey)
}
\arguments{
\item{VRMs}{A list of VRMs as character strings.}

\item{apikey}{Your API key as a character string.}
}
\description{
Download vehicle data from the DVSA MOT API using VRM.
}
\section{Details}{

This function takes a list of vehicle registrations (VRMs) and returns vehicle data from MOT records.
It returns a data frame of those VRMs which were successfully used with the DVSA MOT API

The DVSA MOT API requires a registration.  The function therefore requires the API key provided by the DVSA.
Be aware that the API has usage limits.  The function will therefore limit lists with more than 150,000 VRMS.
}

\examples{
\donttest{
get_MOT_STATS19(VRMs = VRMs, apikey = apikey)
}
}
