% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parquet_directory}
\alias{get_parquet_directory}
\title{Get Parquet directory}
\usage{
get_parquet_directory()
}
\description{
Returns the directory where converted Parquet files are stored.
Defaults to the value of the \code{STATS19_PARQUET_DIRECTORY} environment variable,
or a "parquet" subdirectory within \code{get_data_directory()}.
}
\examples{
get_parquet_directory()
}
