% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulez.R
\name{get_ULEZ}
\alias{get_ULEZ}
\title{Download DVLA-based vehicle data from the TfL API using VRM.}
\usage{
get_ULEZ(vrm)
}
\arguments{
\item{vrm}{A list of VRMs as character strings.}
}
\description{
Download DVLA-based vehicle data from the TfL API using VRM.
}
\section{Details}{

This function takes a a character vector of vehicle registrations (VRMs) and returns DVLA-based vehicle data from TfL's API, included ULEZ eligibility.
It returns a data frame of those VRMs which were successfully used with the TfL API.  ULEZ-exempt vehicles will not have vehicle details returned - they will simply be marked "exempt".

Be aware that the API has usage limits.  The function will therefore limit API calls to 50 per minute.
}

\examples{
\donttest{
vrm = c("1RAC","P1RAC")
if(nchar(apikey) > 0) {
  get_ULEZ(vrm = vrm)
}
}
}
