% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_files}
\alias{locate_files}
\title{Locate a file on disk}
\usage{
locate_files(
  data_dir = get_data_directory(),
  type = NULL,
  years = NULL,
  quiet = FALSE
)
}
\arguments{
\item{data_dir}{Where sets of downloaded data would be found.}

\item{type}{One of 'collision', 'casualty', 'Vehicle'; defaults to 'collision'.}

\item{years}{Single year or vector of years for which data are to be read.}

\item{quiet}{Print out messages (files found)}
}
\description{
Locate a file on disk
}
