% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_one_file}
\alias{locate_one_file}
\title{Pin down a file on disk from parameters.}
\usage{
locate_one_file(
  filename = NULL,
  data_dir = get_data_directory(),
  year = NULL,
  type = NULL
)
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{year}{Single year for which data are to be read.}

\item{type}{One of 'collision', 'casualty', 'Vehicle'; defaults to 'collision'.}
}
\description{
Pin down a file on disk from parameters.
}
\examples{
\donttest{
locate_one_file()
}
}
