% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_boundaries}
\alias{police_boundaries}
\title{Police force boundaries in England (2016)}
\format{An sf data frame}
\description{
See http://geoportal.statistics.gov.uk/
}
\examples{
\donttest{
# Obtained with:
library(geojsonsf)
u = "https://opendata.arcgis.com/datasets/3e5a096a8c7c456fb6d3164a3f44b005_3.geojson"
police_boundaries_wgs = geojsonsf::geojson_sf(u)
names(police_boundaries_wgs)
police_boundaries = st_transform(police_boundaries_wgs, 27700)
names(police_boundaries)
police_boundaries = police_boundaries[c("pfa16cd", "pfa16nm")]
}
}
\keyword{datasets}
