% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/format.R
\docType{data}
\name{stats19_schema}
\alias{stats19_schema}
\alias{read_schema}
\title{Stats19 schema}
\usage{
read_schema(dir = tempdir(),
  filename = "Road-Accident-Safety-Data-Guide.xls", sheet = NULL)
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read in - default values
are those downloaded from the UK Department for Transport (DfT).}

\item{sheet}{integer to be added if you want to download a single sheet}
}
\description{
This function generates the data object `stats19_schema` in a reproducible way
using DfT's schema definition (see function [dl_schema()]).
}
\examples{
\dontrun{
stats19_schema = read_schema()
}

}
\keyword{datasets}
