% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_file}
\alias{select_file}
\title{Interactively select from options}
\usage{
select_file(fnames)
}
\arguments{
\item{fnames}{Character vector of filenames to select from.}
}
\description{
Interactively select from options
}
