% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_data_directory}
\alias{set_data_directory}
\title{Set data download dir}
\usage{
set_data_directory(data_path)
}
\arguments{
\item{data_path}{valid existing path to save downloaded files in.}
}
\description{
Set data download dir
}
