% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parquet_directory}
\alias{set_parquet_directory}
\title{Set Parquet directory}
\usage{
set_parquet_directory(parquet_path)
}
\arguments{
\item{parquet_path}{Valid path to save Parquet files in. Created if it does not exist.}
}
\description{
Sets the \code{STATS19_PARQUET_DIRECTORY} environment variable to specify where
converted Parquet files should be saved and read from.
}
\examples{
\donttest{
set_parquet_directory(tempdir())
}
}
