% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats19_schema}
\alias{stats19_schema}
\alias{stats19_variables}
\title{Stats19 schema and variables}
\description{
\code{stats19_schema} and \code{stats19_variables} contain
metadata on \pkg{stats19} data.
\code{stats19_schema} is a look-up table matching
codes provided in the raw stats19 dataset with
character strings.
}
\note{
The schema data can be (re-)generated using the script in the
\code{data-raw} directory.
}
\section{Known caveat - junction_detail code 0}{

\code{junction_detail} code 0 is labelled "Not at junction or within 20
metres" in the DfT data guide, but from 2015 the same code also covers
roundabouts, mini-roundabouts and slip roads: the guide does not split
these out. \code{junction_detail_historic} still separates them for 2023 and
earlier, but is unknown for every row from 2024 onwards. Use
\code{\link[=junction_status]{junction_status()}} to get a clean junction/not junction classification
instead of filtering on \code{junction_detail} directly. See
\url{https://github.com/ropensci/stats19/issues/328}.
}

\keyword{datasets}
