% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents_sample}
\alias{accidents_sample}
\alias{accidents_sample_raw}
\title{Sample of stats19 data (2017 accidents)}
\format{
A data frame
}
\description{
Sample of stats19 data (2017 accidents)
}
\note{
These were generated using the script in the
\code{data-raw} directory (\code{misc.Rmd} file).
}
\examples{
\donttest{
nrow(accidents_sample_raw)
accidents_sample_raw
}
}
\keyword{datasets}
