% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats19.R
\name{check_input_file}
\alias{check_input_file}
\title{Local helper to be reused.}
\usage{
check_input_file(filename = NULL, type = NULL, data_dir = NULL,
  years = NULL)
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
years determine whether there is a target to read, otherwise disk scan would be needed.}

\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'#'}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{years}{Either a single year or a two year range, defaults to 2 years ago}
}
\description{
Local helper to be reused.
}
