% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year}
\alias{check_year}
\title{This function does two things:
\enumerate{
\item is used to check if there is an overlapping of files with
multiple years. The matching between the years and the files works as follows:
1979 ... 2004 ---> 1979 - 2004
2005 ... 2008 ---> 2005 - 2014
2009          ---> 2009
2010          ---> 2010
2011          ---> 2011
...
2018          ---> 2018
\item it also does the sanity checking of the year(s) given
}}
\usage{
check_year(year)
}
\arguments{
\item{year}{Single year for which file is to be downloaded.}
}
\description{
This function does two things:
\enumerate{
\item is used to check if there is an overlapping of files with
multiple years. The matching between the years and the files works as follows:
1979 ... 2004 ---> 1979 - 2004
2005 ... 2008 ---> 2005 - 2014
2009          ---> 2009
2010          ---> 2010
2011          ---> 2011
...
2018          ---> 2018
\item it also does the sanity checking of the year(s) given
}
}
\examples{
# check_year("2018") # fails
# check_year(2017)
# check_year(2006)
# check_year(1985)
}
