% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{dl_schema}
\alias{dl_schema}
\title{Download stats19 schema}
\usage{
dl_schema(data_dir = tempdir())
}
\arguments{
\item{data_dir}{Location to download, defaults to `tempdir()`}
}
\description{
This function downloads an excel spreadsheet containing variable names and categories
}
\examples{
\dontrun{
dl_schema()
}
}
