% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{dl_stats19}
\alias{dl_stats19}
\title{Download STATS19 data for a year}
\usage{
dl_stats19(
  year = NULL,
  type = NULL,
  data_dir = get_data_directory(),
  file_name = NULL,
  ask = FALSE,
  silent = FALSE,
  timeout = 600
)
}
\arguments{
\item{year}{Single year for which data are to be read}

\item{type}{One of 'collision', 'casualty', 'Vehicle'; defaults to 'collision'.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{file_name}{Character string of a specific STATS19 CSV filename to
download/read. If \code{NULL}, filenames are inferred from \code{year} and \code{type}.}

\item{ask}{Should you be asked whether or not to download the files? \code{TRUE} by default.}

\item{silent}{Boolean. If \code{FALSE} (default value), display useful progress
messages on the screen.}

\item{timeout}{Timeout in seconds for the download if current option is less than
this value. Defaults to 600 (10 minutes).}
}
\description{
Download STATS19 data for a year
}
\examples{
\donttest{
if (curl::has_internet()) {
  # type by default is collisions table
  dl_stats19(year = 2022)
}
}
}
