% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_stats19.R
\name{dl_stats19}
\alias{dl_stats19}
\title{Download Stats19 data for a year or range of two years.}
\usage{
dl_stats19(year = NULL, type = NULL, data_dir = tempdir(),
  file_name = NULL)
}
\arguments{
\item{year}{Single year for which file is to be downloaded}

\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'#'}

\item{data_dir}{Parent directory for all downloaded files. Defaults to `tempdir()`}

\item{file_name}{The file name to download, above two will be ignore.}
}
\description{
Download Stats19 data for a year or range of two years.
}
\section{Details}{

This convenience function downloads and unzips UK road crash data.
It results in unzipped .csv files that are put
in the temporary directory specified by `tempdir()`.

If you move the files into your current working directory
and run the same command again, the file will not
downloaded.

The `dl_*` functions can download many MB of data so ensure you
have a sufficient internet access and hard disk space.
}

\examples{
\dontrun{
dl_stats19(year = 2017) # interactively select files...

# now you can read-in the data
dl_stats19(year = 2009)
dl_stats19(year = 2009, type = "casualties")
dl_stats19(year = 1985)
}
}
