% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{file_names}
\alias{file_names}
\title{File names for easy access}
\format{A named list}
\description{
URL encoded file names. Generated as follows:
Visit URL from Chrome/Firefox
https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data
}
\details{
Enter the browser JavaScript console and run:
filenames = []
list = document.getElementsByClassName("dgu-datafile")
for (var i = 0; i < list.length; i++) {
 filenames.push(list[i].firstElementChild.children[0].href.split("/")[4]);
}
now copy to clipboard
copy(filenames)

paste into Rstudio and create a vector (file_names)
then
file_names = setNames(as.list(file_names), file_names)
usethis::use_data(file_names)
}
\keyword{datasets}
