% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{file_names}
\alias{file_names}
\title{stats19 file names for easy access}
\format{A named list}
\description{
URL decoded file names. Currently there are 52 file names
released by the DfT (Department for Transport) and the details include
how these were obtained and would be kept up to date.
}
\details{
The file names were generated as follows:
}
\examples{
\dontrun{
library(rvest)
library(stringr)
u = paste0(
  "https://data.gov.uk/dataset/",
  "cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
page = read_html(u)

r = page \%>\%
  html_nodes("a") \%>\%       # find all links
  html_attr("href") \%>\%     # get the url
  str_subset("\\\\.zip")

dr = c()
for(i in 1:length(r)) {
  dr[i] = sub("http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/",
              "", URLdecode(r[i]))
  dr[i] = sub("http://data.dft.gov.uk/road-accidents-safety-data/",
              "", dr[i])
}
file_names = setNames(as.list(file_names), file_names)
usethis::use_data(file_names)
}
}
\keyword{datasets}
