% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file_name}
\alias{find_file_name}
\title{Find file names within stats19::file_names.}
\usage{
find_file_name(years = NULL, type = NULL)
}
\arguments{
\item{years}{Year for which data are to be found}

\item{type}{One of 'collisions', 'casualty' or
'vehicles' ignores case.}
}
\description{
Find file names within stats19::file_names.
}
\examples{
find_file_name(2016)
}
