% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file_name}
\alias{find_file_name}
\title{Find a file name within stats19::file_names.}
\usage{
find_file_name(years = "2016", type = "")
}
\arguments{
\item{years}{Provide a year, part of it or maximum range of two years. Defaults to two years ago.}

\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}
}
\description{
Currently, there are 52 file names to download data from.
}
