% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file_name}
\alias{find_file_name}
\title{Find file names within stats19::file_names.}
\usage{
find_file_name(years = NULL, type = NULL)
}
\arguments{
\item{years}{Year for which data are to be found}

\item{type}{One of 'collisions', 'casualty' or
'vehicles' ignores case.}
}
\description{
Currently, there are 52 file names to download/read data from.
}
\examples{
find_file_name(2016)
find_file_name(2016, type = "collision")
find_file_name(1985, type = "collision")
find_file_name(type = "cas")
find_file_name(type = "collision")
find_file_name(2016:2022) # warning when multiple years requested
}
