% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file_name}
\alias{find_file_name}
\title{Find file names within stats19::file_names.}
\usage{
find_file_name(years = NULL, type = NULL)
}
\arguments{
\item{years}{Years for which data are to be found}

\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}
}
\description{
Currently, there are 52 file names to download/read data from.
}
\examples{
find_file_name(2016)
find_file_name(2016, type = "Accidents")
find_file_name(1985, type = "Accidents")
find_file_name(type = "cas")
find_file_name(type = "accid")
find_file_name(2006)
find_file_name(2016:2017)
}
