% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_accidents}
\alias{format_accidents}
\title{Format UK 'Stats19' road traffic casualty data}
\usage{
format_accidents()
}
\arguments{
\item{ve}{Dataframe representing the raw Stats19 data read-in with `read_csv()`.}
}
\description{
Format UK 'Stats19' road traffic casualty data
}
\section{Details}{

This is a helper function to format raw stats19 data
}

\examples{
\dontrun{
ve <- format_stats19_2005_2014_ve(ve)
}
}
