% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_accidents}
\alias{format_accidents}
\title{Format stats19 'accidents' data}
\usage{
format_accidents(x, factorize = FALSE)
}
\arguments{
\item{x}{Data frame created with `read_accidents()`}

\item{factorize}{Should some results be returned as factors? `FALSE` by default}
}
\description{
Format stats19 'accidents' data
}
\section{Details}{

This is a helper function to format raw stats19 data
}

\examples{
\dontrun{
x = read_accidents()
sapply(x, class)
table(x$Accident_Severity)
crashes = format_accidents(x)
sapply(crashes, class)
table(crashes$accident_severity)
}
}
