% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_accidents}
\alias{format_accidents}
\title{Format STATS19 'accidents' data}
\usage{
format_accidents(x)
}
\arguments{
\item{x}{Data frame created with \code{read_accidents()}}
}
\description{
Format STATS19 'accidents' data
}
\section{Details}{

This is a helper function to format raw STATS19 data
}

\examples{
\donttest{
dl_stats19(year = 2017, type = "accident")
x = read_accidents(year = 2017)
crashes = format_accidents(x)
}
}
