% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_accidents}
\alias{format_accidents}
\title{Format STATS19 'accidents' data}
\usage{
format_accidents(x)
}
\arguments{
\item{x}{Data frame created with \code{read_accidents()}}
}
\description{
Format STATS19 'accidents' data
}
\section{Details}{

This is a helper function to format raw STATS19 data
}

\examples{
\donttest{
if(curl::has_internet()) {
dl_stats19(year = 2017, type = "accident")
x = read_accidents(year = 2017, format = FALSE)
if(nrow(x) > 0) {
x[1:3, 1:12]
crashes = format_accidents(x)
crashes[1:3, 1:12]
summary(crashes$datetime)
}
}
}
}
