% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_casualties}
\alias{format_casualties}
\title{Format STATS19 casualties}
\usage{
format_casualties(x)
}
\arguments{
\item{x}{Data frame created with \code{read_casualties()}}
}
\description{
Format STATS19 casualties
}
\section{Details}{

This function formats raw STATS19 data
}

\examples{
\donttest{
if(curl::has_internet()) {
dl_stats19(year = 2017, type = "casualty")
x = read_casualties(year = 2017)
casualties = format_casualties(x)
}
}
}
