% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats19.R
\name{format_stats19_2005_2014_ac}
\alias{format_stats19_2005_2014_ac}
\title{Format UK 'Stats19' road traffic casualty data}
\usage{
format_stats19_2005_2014_ac(ac)
}
\arguments{
\item{ac}{Dataframe representing the raw Stats19 data read-in with `read_csv()`.}
}
\description{
Format UK 'Stats19' road traffic casualty data
}
\section{Details}{

This is a helper function to format raw stats19 data
}

\examples{
\dontrun{
ac <- format_stats19_2005_2014_ac(ac)
}
}
