% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_vehicles}
\alias{format_vehicles}
\title{Format stats19 vehicles data}
\usage{
format_vehicles(x)
}
\arguments{
\item{x}{Data frame created with `read_vehicles()`}
}
\description{
Format stats19 vehicles data
}
\section{Details}{

This function formats raw stats19 data
}

\examples{
\dontrun{
dl_stats19(year = 2017, type = "vehicles")
x = read_vehicles(year = 2017)
vehicles = format_vehicles(x)
}
}
