% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_stats19}
\alias{get_stats19}
\title{Download, read and format STATS19 data in one function.}
\usage{
get_stats19(year = NULL, type = NULL, data_dir = tempdir(),
  file_name = NULL)
}
\arguments{
\item{year}{Single year for which file is to be downloaded.}

\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'.
Or any variation of to search the file names with such as "acc" or "accid".}

\item{data_dir}{Parent directory for all downloaded files. Defaults to `tempdir()`.}

\item{file_name}{The file name (DfT named) to download.}
}
\description{
Download, read and format STATS19 data in one function.
}
\section{Details}{

This function utilizes `dl_stats19` and `read_*` functions
and retuns a df. The file downloaded would be for a specific year (e.g 2017).

As this function uses `dl_stats19` function, it can download
many MB of data so ensure you have a sufficient disk space.
}

\examples{
\dontrun{
get_stats19(year = 2017)
get_stats19(year = 2009)
}
}
\seealso{
\code{\link{dl_stats19}}

\code{\link{read_accidents}}
}
