% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Build the API endpoints}
\usage{
get_url(file_name = "")
}
\arguments{
\item{file_name}{(Optional) to add to the url returned.}
}
\description{
Build the API endpoints
}
\examples{
\dontrun{
get_url()
}
}
