% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Convert file names to urls}
\usage{
get_url(
  file_name = "",
  domain = "https://data.dft.gov.uk",
  directory = "road-accidents-safety-data"
)
}
\arguments{
\item{file_name}{Optional file name to add to the url returned (empty by default)}

\item{domain}{The domain from where the data will be downloaded}

\item{directory}{The subdirectory of the url}
}
\description{
Convert file names to urls
}
\details{
This function returns urls that allow data to be downloaded from the pages:

https://data.dft.gov.uk/road-accidents-safety-data/RoadSafetyData_2015.zip

Last updated: October 2020.
Files available from the s3 url in the default \code{domain} argument.
}
\examples{
# get_url(find_file_name(1985))
}
