% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_files}
\alias{locate_files}
\title{Locate a file on disk}
\usage{
locate_files(data_dir = tempdir(), type = "Accidents", years = NULL,
  quiet = FALSE)
}
\arguments{
\item{data_dir}{Super directory where dataset(s) were first downloaded to.}

\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}

\item{quiet}{Print out messages (files found)}

\item{year}{Years for which data are to be found}
}
\value{
Depending on @param return: full path of a single file found, list of directories
where data from DfT (stats19::filenames) have been downloaded to or NULL.
}
\description{
Helper function to locate files. Given below params, the function
returns 0 or more files found at location/names given.
}
\examples{
locate_files(years = 2016)
}
