% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_one_file}
\alias{locate_one_file}
\title{Pin down a file on disk from four parameters.}
\usage{
locate_one_file(filename = "", data_dir = tempdir(), years = "",
  type = "accidents")
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
years determin whether there is a target to read, otherwise disk scan would be needed.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{years}{Either a single year or a two year range, defaults to 2 years ago}

\item{type}{One of: 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}
}
\value{
One of: path for one file, a message `More than one file found` or NULL
}
\description{
Pin down a file on disk from four parameters.
}
\examples{
\dontrun{
locate_one_file()
locate_one_file(filename = "Cas.csv")
}
}
