% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_one_file}
\alias{locate_one_file}
\title{Pin down a file on disk from four parameters.}
\usage{
locate_one_file(
  filename = NULL,
  data_dir = get_data_directory(),
  year = NULL,
  type = NULL
)
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read, if this
is given, type and years determine whether there is a target to read,
otherwise disk scan would be needed.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{year}{Single year for which file is to be found.}

\item{type}{One of: 'Accidents', 'Casualties', 'Vehicles'; ignores case.}
}
\value{
One of: path for one file, a message \verb{More than one file found} or error if none found.
}
\description{
Pin down a file on disk from four parameters.
}
\examples{
\donttest{
locate_one_file()
locate_one_file(filename = "Cas.csv")
}
}
