% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_boundaries}
\alias{police_boundaries}
\title{Police force boundaries in England (2016)}
\format{An sf data frame}
\description{
See http://geoportal.statistics.gov.uk/
}
\note{
These were generated using the script in the
\code{data-raw} directory (\code{misc.Rmd} file).
}
\examples{
\donttest{
nrow(police_boundaries)
police_boundaries[police_boundaries$pfa16nm == "West Yorkshire", ]
}
}
\keyword{datasets}
