% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_accidents}
\alias{read_accidents}
\title{Import and Stats19 data on road traffic casualties}
\usage{
read_accidents(filename = "", data_dir = tempdir(), year = NULL)
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
years determine whether there is a target to read, otherwise disk scan would be needed.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{year}{Single year for which data are to be read}
}
\description{
Import and Stats19 data on road traffic casualties
}
\section{Details}{

This is a wrapper function to access and load stats 19 data in a user-friendly way.
The function returns a data frame, in which each record is a reported incident in the
stats19 dataset.
}

\examples{
\dontrun{
dl_stats19(year = 2017, type = "Accidents")
ac = read_accidents(year = 2017)
}
}
