% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats19.R
\name{read_accidents}
\alias{read_accidents}
\title{Import and Stats19 data on road traffic casualties}
\usage{
read_accidents(filename = "", data_dir = tempdir(), years = "")
}
\arguments{
\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
years determine whether there is a target to read, otherwise disk scan would be needed.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{years}{Either a single year or a two year range, defaults to 2 years ago}
}
\description{
Import and Stats19 data on road traffic casualties
}
\section{Details}{

This is a wrapper function to access and load stats 19 data in a user-friendly way.
The function returns a data frame, in which each record is a reported incident in the
stats19 dataset.
}

\examples{
\dontrun{
ac = read_accidents()
}
}
