% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_casualties}
\alias{read_casualties}
\title{Read in stats19 road safety data from .csv files downloaded.}
\usage{
read_casualties(year = NULL, filename = "", data_dir = tempdir(),
  format = TRUE)
}
\arguments{
\item{year}{Single year for which data are to be read}

\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
years determine whether there is a target to read, otherwise disk scan would be needed.}

\item{data_dir}{Where sets of downloaded data would be found.}

\item{format}{Switch to return raw read from file, default is `FALSE`.}
}
\description{
Read in stats19 road safety data from .csv files downloaded.
}
\section{Details}{

The function returns a data frame, in which each record is a reported casualty
in the stats19 dataset.
}

\examples{
\dontrun{
dl_stats19(years = 2017, type = "casualties")
casualties = read_casualties()
}
}
