% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{read_schema}
\alias{read_schema}
\title{Load stats19 schema}
\usage{
read_schema(data_dir = tempdir(),
  filename = "Road-Accident-Safety-Data-Guide.xls", sheet = NULL)
}
\arguments{
\item{data_dir}{Character string representing where the data is stored.
Default value assumes in recent session `tempdir` there is
a `dftRoadSafetyData_Accidents_2017` directory.}

\item{filename}{Character string of the filename of the .csv to read in - default values
are those downloaded from the UK Department for Transport (DfT).
Default value is `Acc.csv` under above default `data_dir`.}

\item{sheet}{integer to be added if you want to download a single sheet}
}
\description{
This function generates the data object `stats19_schema` in a reproducible way
using DfT's schema definition (see function [dl_schema()]).
}
\details{
The function also generates `stats19_variables`
(see the function's source code for details).
}
\examples{
\dontrun{
stats19_schema = read_schema()
}
}
