% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{read_schema}
\alias{read_schema}
\title{Load stats19 schema}
\usage{
read_schema(data_dir = tempdir(),
  filename = "Road-Accident-Safety-Data-Guide.xls", sheet = NULL)
}
\arguments{
\item{data_dir}{Where sets of downloaded data would be found.}

\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
years determine whether there is a target to read, otherwise disk scan would be needed.}

\item{sheet}{integer to be added if you want to download a single sheet}
}
\description{
This function generates the data object `stats19_schema` in a reproducible way
using DfT's schema definition (see function [dl_schema()]).
}
\details{
The function also generates `stats19_variables`
(see the function's source code for details).
}
\examples{
\dontrun{
stats19_schema = read_schema()
}
}
