% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats19.R
\name{read_vehicles}
\alias{read_vehicles}
\title{Import and Stats19 data on vehicles}
\usage{
read_vehicles(data_dir = file.path(tempdir(),
  "dftRoadSafetyData_Vehicles_2017"), filename = "Veh.csv")
}
\arguments{
\item{data_dir}{Character string representing where the data is stored.
If empty, R will attempt to download and unzip the data for you.}

\item{filename}{Character string of the filename of the .csv to read in - default value
is `Veh.csv` from default data_dir}
}
\description{
Import and Stats19 data on vehicles
}
\section{Details}{

The function returns a data frame, in which each record is a reported vehicle in the
stats19 dataset for the data_dir and filename provided.
}

\examples{
\dontrun{
ve = read_vehicles()
}
}
