% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_data_directory}
\alias{set_data_directory}
\title{Set data download dir}
\usage{
set_data_directory(data_path)
}
\arguments{
\item{data_path}{valid existing path to save downloaded files in.}
}
\description{
Handy function to manage \code{stats19} package underlying environment
variable. If run interactively it makes sure user does not change
directory by mistatke.
}
\examples{
# set_data_directory("MY_PATH")
}
