% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stats19_schema}
\alias{stats19_schema}
\alias{stats19_variables}
\title{Stats19 schema and variables}
\description{
`stats19_schema` and `stats19_variables` contain
metadata on stats19 data.
`stats19_schema` is a look-up table matching
codes provided in the raw stats19 dataset with
character strings.
}
\keyword{datasets}
