% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicles_sample}
\alias{vehicles_sample}
\alias{vehicles_sample_raw}
\title{Sample of stats19 data (2017 vehicles)}
\format{
A data frame
}
\description{
Sample of stats19 data (2017 vehicles)
}
\note{
These were generated using the script in the
\code{data-raw} directory (\code{misc.Rmd} file).
}
\examples{
\donttest{
nrow(vehicles_sample_raw)
vehicles_sample_raw
}
}
\keyword{datasets}
